requirements.txt
Flask
Pillow
praw
python-dotenv
requests
rich
selenium
line_bot_sdk
pymongo[srv]
fuzzywuzzy
binusmayapy
python-Levenshtein
openai